requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
opencv-python>=4.8.1.78
watchdog>=3.0.0
numpy>=1.26.0
pymupdf>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pywin32>=306
pyodbc>=5.0.1
pytest>=7.4.0
hypothesis>=6.82.0
torch
torchvision
scikit-learn
spacy
openai
python-dotenv
httpx
fastapi
uvicorn

